Optionally add the time offset that GENIE provides for travel from flux window to interaction vertex (and depending on flux driver possibly from proton-on-target to flux window).
No longer try funky things with individual particles that aren't status 0 or 1
in regards to whether to put the positon in the lab frame; the Gvtx data is
safely storing the relative-to-the-struck-nucleus position we need for
recovering the GENIE event record used in reweighting.

Units for an overall shift are now specified (cm,ns) that MCTruth uses.

Remove some stale code (commented out) and some wrong-headed attempt to
deal with what looked like bad Gvtx positions.